Separa componente PokemonCarta de ListaPokemon
