+++ src/components/Inicio/ListadoPokemon.jsx
@@ -1,9 +1,7 @@
 import React from "react";
-import { Link } from "react-router-dom";
 import styled from "@emotion/styled";
-import { numeroTresCifras } from "../../utils/numeroTresCifras";
-import { primerLetraMayus } from "../../utils/primerLetraMayus";
 import { obtenerIDdeURL } from "../../utils/obtenerIDdeURL";
+import { CartaPokemon } from "./CartaPokemon";
 
 const ContenedorListado = styled.div`
   margin: 1em;
@@ -13,41 +11,7 @@ const ContenedorListado = styled.div`
   min-height: 320px;
 `;
 
-const ContenedorCartaPokemon = styled.div`
-  margin: 2.5px;
-  background-color: gray;
-  height: 160px;
-  width: 160px;
-  border: 1px solid white;
-`;
-
-const NombrePokemon = styled.p`
-  font-weight: bold;
-`;
-
-const ImagenPokemon = styled.img`
-  height: 120px;
-`;
-
-const CartaPokemon = ({ indicePokemon, pokemon }) => {
-  const pokemonFormateado = primerLetraMayus(pokemon);
-  const indicePokemonFormateado = numeroTresCifras(indicePokemon);
-  return (
-    <ContenedorCartaPokemon>
-      <Link style={{ color: "black" }} to={`/pokemon/${pokemon}`}>
-        <ImagenPokemon
-          src={`https://assets.pokemon.com/assets/cms2/img/pokedex/full/${indicePokemonFormateado}.png`}
-          alt={pokemon}
-        ></ImagenPokemon>
-        <NombrePokemon>
-          {`#${indicePokemon} ${pokemonFormateado}`}
-        </NombrePokemon>
-      </Link>
-    </ContenedorCartaPokemon>
-  );
-};
-
-const ListadoPokemon = ({ listaPokemones }) => {
+export const ListadoPokemon = ({ listaPokemones }) => {
   return (
     <ContenedorListado>
       {listaPokemones != null &&
@@ -61,5 +25,3 @@ const ListadoPokemon = ({ listaPokemones }) => {
     </ContenedorListado>
   );
 };
-
-export default ListadoPokemon;
